parkhang/requirements/base.txt
# Requirements for all environments

# General
requests==2.27.1

# Django
# -----------------------------------------------------------------------
Django==3.2.12
django-environ==0.8.1  # https://github.com/joke2k/django-environ
django-webpack-loader==1.4.1
whitenoise==5.3.0  # https://github.com/evansd/whitenoise


# Django REST Framework
djangorestframework==3.13.1
